fix tests
by adding profile to MockTool
